# optimize yield performance