```
Calculating -------------------------------------
                              master        ours
                 vm_block    29.130M     33.747M i/s -     30.000M times in 1.029875s 0.888962s
  block_handler_type_iseq      9.223      10.228 i/s -       1.000 times in 0.108424s 0.097771s
block_handler_type_symbol      5.367       5.724 i/s -       1.000 times in 0.186336s 0.174703s
  block_handler_type_proc      8.908       9.424 i/s -       1.000 times in 0.112259s 0.106109s
            vm_blockparam    30.180M     30.723M i/s -     30.000M times in 0.994049s 0.976456s
       vm_blockparam_call    16.461M     17.924M i/s -     30.000M times in 1.822495s 1.673779s
       vm_blockparam_pass    14.157M     15.379M i/s -     30.000M times in 2.119050s 1.950723s
      vm_blockparam_yield    18.791M     20.141M i/s -     30.000M times in 1.596481s 1.489500s

Comparison:
                              vm_block
                     ours:  33747230.4 i/s
                   master:  29129752.3 i/s - 1.16x  slower

               block_handler_type_iseq
                     ours:        10.2 i/s
                   master:         9.2 i/s - 1.11x  slower

             block_handler_type_symbol
                     ours:         5.7 i/s
                   master:         5.4 i/s - 1.07x  slower

               block_handler_type_proc
                     ours:         9.4 i/s
                   master:         8.9 i/s - 1.06x  slower

                         vm_blockparam
                     ours:  30723358.1 i/s
                   master:  30179590.8 i/s - 1.02x  slower

                    vm_blockparam_call
                     ours:  17923514.4 i/s
                   master:  16460951.4 i/s - 1.09x  slower

                    vm_blockparam_pass
                     ours:  15378917.1 i/s
                   master:  14157284.2 i/s - 1.09x  slower

                   vm_blockparam_yield
                     ours:  20140982.6 i/s
                   master:  18791326.3 i/s - 1.07x  slower

```