@ko1 2020-05-28
could you explain the idea?