@shyouhei 2020-05-28
There is in  fact no such thing like a big idea behind this pull request... It is a set of random improvements.  However I guess these commits work:

- https://github.com/ruby/ruby/pull/3152/commits/f8a8bd52f7e32ec37882659813021fa921efefde ... This commit changes `vm_invoke_block`'s children so that they can be tail-call optimized.  Function calls in `vm_invoke_block` were `CALL` instructions before, but now they are `JUMP` instructions.

- https://github.com/ruby/ruby/pull/3152/commits/0ecde78b60a2b602c4e16414d8b3f10b4a064e7b ...  It reduces `vm_exec_core`.

    before:
    ```
    + vm_invokeblock_i
      + vm_invoke_block
        + vm_invoke_symbol_block
          + vm_yield_with_symbol
            + rb_sym_proc_call
              + rb_funcall_with_block_kw
                + rb_call
                  + rb_call0
                    + rb_vm_call_kw
                      + rb_vm_call0
                        + vm_call0_body
                          + vm_exec
    ```

    after:
    ```
    + vm_invokeblock_i
      + vm_invoke_block
        + vm_invoke_symbol_block
          + vm_call_opt_send
            + vm_call_method
              + vm_call_method_each_type
                + vm_call_iseq_setup
    ```