@nobu 2020-05-28
I'm uncertain  if these tail call recursions are optimized as JUMP always.