@ko1 2020-05-29
Thank you for explanation.
I understand we need to check per commit (i can't get the idea by checking the whole diff).
I'll check them.