@mame 2020-05-29
I applied optcarrot's [tools/statistic-test.rb](https://github.com/mame/optcarrot/blob/master/tools/statistic-test.rb) and it cannot see significant differences between the current master and this PR.

```
$ ruby tools/statistic-test.rb
Install the spreadsheet gem version ~>1.1.1 for using spreadsheet functions.
Install the mechanize gem version ~>2.7.5 for using mechanize functions.
Doing `require 'backports'` is deprecated and will not load any backport in the next major release.
Require just the needed backports instead, or 'backports/latest'.
new     master (in fps)
34.247  35.355
34.359  33.143
35.201  31.369
34.085  34.646
33.836  33.423
33.109  34.418
34.042  33.264
32.903  33.782
33.717  35.035
31.475  34.203
34.182  34.062
33.821  33.307
34.743  33.416
34.085  32.935
33.622  33.496
33.829  31.945
33.636  34.195
33.184  34.491
33.345  33.164
31.226  22.225
32.762  32.780
34.041  32.678
33.082  34.808
32.722  34.501
32.163  32.786
33.844  32.769
30.992  33.938
31.847  33.819
31.790  34.358
33.614  33.386
/home/mame/local/lib/ruby/gems/2.7.0/gems/distribution-0.7.3/lib/distribution/math_extension/incomplete_beta.rb:14: warning: constant ::Fixnum is deprecated
/home/mame/local/lib/ruby/gems/2.7.0/gems/distribution-0.7.3/lib/distribution/math_extension/incomplete_beta.rb:14: warning: constant ::Fixnum is deprecated
= Two Sample T Test
  Mean and standard deviation
+----------+---------+--------+----+
| Variable |  mean   |   sd   | n  |
+----------+---------+--------+----+
|          | 33.3168 | 1.0499 | 30 |
|          | 33.2565 | 2.2687 | 30 |
+----------+---------+--------+----+
  Levene test for equality of variances : F(1, 58) = 0.5617 , p = 0.4566
  T statistics
+--------------------+--------+---------+----------------+
|        Type        |   t    |   df    | p (both tails) |
+--------------------+--------+---------+----------------+
| Equal variance     | 0.1320 | 58      | 0.8954         |
| Non equal variance | 0.1320 | 40.8771 | 0.8956         |
+--------------------+--------+---------+----------------+
  Effect size
+-------+--------+
| x1-x2 | 0.0603 |
| d     | 0.0476 |
+-------+--------+
p-value is 0.896 >= 0.05; There is NO significant differences
```