@shyouhei 2020-05-29
> @mame san pointed out that it requires stackoverflow check.

That part is updated by https://github.com/ruby/ruby/pull/3152/commits/4fd5998752ce9543b9c701774c07d2a7e4826b35.  I believe there is no need to check stack overflow in the end.