@ko1 2020-05-29
(not sure why it is not in a thread...)

> > @mame san pointed out that it requires stackoverflow check.
> 
> That part is updated by [4fd5998](https://github.com/ruby/ruby/commit/4fd5998752ce9543b9c701774c07d2a7e4826b35). I believe there is no need to check stack overflow in the end.

sorry i can't find the reason why it is not needed. I may missed it could you point out the reason?