@ko1 2020-05-29
I understand the big impact of this PR is simplify the symbol block invocation because we don't need to call `rb_call()`. It is nice.
Also I wonder that other refactoring contribute the general block invocation. No objection but it will be helpful to show the gcc version benchmark results.