@shyouhei 2020-05-29
> sorry i can't find the reason why it is not needed. I may missed it could you point out the reason?

- https://github.com/ruby/ruby/pull/3152/commits/0ecde78b60a2b602c4e16414d8b3f10b4a064e7b : When `&:sym` is yielded, that symbol is pushed onto the VM stack (can cause stack overflow)
- https://github.com/ruby/ruby/pull/3152/commits/4fd5998752ce9543b9c701774c07d2a7e4826b35 : Instead, the symbol is passed via a C function argument.  In case of `&:sym`, VM stack is unmodified (because the symbol is already externed, `if (UNLIKELY(!mid))` branch never happen).