@ko1 2020-05-29
> In case of &:sym, VM stack is unmodified. Because the symbol is already interned, if (UNLIKELY(!mid)) branch never happen.

Oh. I missed it. Where does it intened?