@shyouhei 2020-05-29
> Oh. I missed it. Where does it intened?

… Ahh. I was wrong. I thought block_handler_type_symbol can only hold interned symbol objects. However there are chances for those symbols to be not pinned yet.

Yes, we need stack oveflow checks.