@shyouhei 2020-05-29
BTW my compiler is clang-11

```
zsh % clang-11 --version
clang version 11.0.0-++20200523063809+304b0ed4039-1~exp1~20200523164413.1701
Target: x86_64-pc-linux-gnu
Thread model: posix
InstalledDir: /usr/bin
```