@nobu 2020-05-30
> > I'm uncertain if these tail call recursions are optimized as JUMP always.
> 
> Of course not. They are simple normal function calls on `-O0`.
> 
> But do you have any example when C guarantees something to be compiled into something? I guess even `register` or `inline` can be ignored.

I meant that unoptimized recursion could result in a stack overflow.
A loop, by `goto` or by `while` etc, would not usually.