@shyouhei 2020-05-31
> > But do you have any example when C guarantees something to be compiled into something? I guess even `register` or `inline` can be ignored.
> 
> I meant that unoptimized recursion could result in a stack overflow.
> A loop, by `goto` or by `while` etc, would not usually.

You mean `vm_invoke_block` can recur infinitely?  Can you tell us how?

And if that happens, is infinite loop forever with no chance to `RUBY_VM_CHECK_INTS` really a better idea than SEGV?