@nobu 2020-05-31
> You mean `vm_invoke_block` can recur infinitely? Can you tell us how?

I don't think it can be **infinite**, but I'm concerned that a long chain might be possible by repeating visibility changes of an inherited method.