@shyouhei 2020-05-31
@nobu Really sorry that I still don't get it.  I think I am missing something.

- The recursion only happens in case of `block_handler_type_proc`.  Right?
- Does it have anything to do with "visibility changes of an inherited method"?